requirements.txt
numpy
scipy
matplotlib
Pillow # For image processing
tqdm  # For progress bars
